Create helpers for the bot info command
